CreamJob          : now memorizes and serializes the delegation expiration time
DelegationManager : now can use also a new method to create the delegation key; the old one is creating that key as SHA1 digest of entire certificate, the new one is just use the DNFQAN from the certificate as key. The DelegationManager::delegate now also returns the expiration time of the created/recycled delegation ID.

iceCommandSubmit  : modified to memorize in the CreamJob also the new delegation's expiration time after. Not using yet the new delegation method.